Customize kopf annotations, fix rtslib dbroot, add raid1 option in LVM, & other cleanup

+++ handler.py
@@ -1,8 +1,8 @@
-from contextlib import suppress
 import logging
 import os
 
 import kopf
+from kopf import Spec, BodyEssence, Meta
 import kubernetes
 
 import config
@@ -29,7 +29,9 @@ def api_login(**kwargs):
 
 def validate_and_register_storage_class(name, sc):
     sc_type = sc.parameters.get("type", "")
-    sc_nodes = sc.parameters.get("nodes", "").split(",")
+    sc_nodes = sc.parameters.get("nodes", "")
+    if sc_nodes:
+        sc_nodes = sc_nodes.split(",") 
 
     if sc_type.lower() not in [Constants.VOLUME_TYPE_ISCSI, Constants.VOLUME_TYPE_NFS, Constants.VOLUME_TYPE_SHARED]:
         logger.error(f"Unrecognized LabDisk volume type: {sc_type.lower()}")
@@ -46,7 +48,7 @@ def validate_and_register_storage_class(name, sc):
         logger.warning(f"Ignoring storage class '{name}' with type '{sc_type}' because the subsystem that handles it is not enabled.")
         return
 
-    if config.get().current_node_name not in sc_nodes:
+    if len(sc_nodes) > 0 and config.get().current_node_name not in sc_nodes:
         logger.info(f"Ignoring storage class '{name}' because it does not apply to this node.")
         return
 
@@ -60,7 +62,13 @@ def new_storageclass(name, **kwargs):
     validate_and_register_storage_class(name, sc)
 
 @kopf.on.startup()
-async def operator_startup(**kwargs):
+async def operator_startup(settings: kopf.OperatorSettings, **kwargs):
+
+    # overwrite default persistence settings
+    settings.persistence.finalizer = Constants.PVC_FINALIZER_KEY
+    settings.persistence.progress_storage = kopf.AnnotationsProgressStorage(prefix=Constants.PERSISTENCE_ANNOTATION_KEY_PREFIX)
+    settings.persistence.diffbase_storage = kopf.AnnotationsDiffBaseStorage(prefix=Constants.PERSISTENCE_ANNOTATION_KEY_PREFIX)
+
     storage_api = kubernetes.client.StorageV1Api()
     storage_classes = storage_api.list_storage_class()
     for sc in storage_classes.items:
@@ -84,11 +92,10 @@ async def operator_startup(**kwargs):
     else:
         logger.info("Individual volume subsystem will be disabled.")
 
-
 @kopf.on.resume("persistentvolume")
-def register_existing_volumes(spec, meta, **kwargs):
+def register_existing_volumes(spec: Spec, meta: Meta, **kwargs):
     storage_class = spec["storageClassName"]
-    pv_name = meta["name"]
+    pv_name = meta.name
 
     # make sure it is a storage class that we manage
     if storage_class not in registered_storage_classes:
@@ -96,7 +103,7 @@ def register_existing_volumes(spec, meta, **kwargs):
         return
 
     # only process volumes on this node
-    asigned_node = meta["annotations"].get(Constants.PV_NODE_ANNOTATION_KEY, "")
+    asigned_node = meta.annotations.get(Constants.PV_ASSIGNED_NODE_ANNOTATION_KEY, "")
     if asigned_node != config.get().current_node_name:
         logger.debug(f"Not registering volume {pv_name} because it is served by a different node")
         return
@@ -136,7 +143,7 @@ def get_storage_class_params(name):
 
     return params
 
-def validate_pvc_spec(spec, meta, update=False):
+def validate_pvc_spec(spec: Spec, meta: Meta, update=False):
     spec = dict(spec)
     storage_class_params = get_storage_class_params(spec["storageClassName"])
     if storage_class_params["type"] != Constants.VOLUME_TYPE_SHARED and ("ReadWriteMany" in spec["accessModes"] or "ReadOnlyMany" in spec["accessModes"]):
@@ -148,12 +155,12 @@ def validate_pvc_spec(spec, meta, update=False):
     return storage_class_params
 
 @kopf.on.create("persistentvolumeclaim")
-def create_volume(meta, spec, **kwargs):
-    storage_class_params = validate_pvc_spec(spec, meta)
-
+def create_volume(meta: Meta, spec: Spec, **kwargs):
     if spec["storageClassName"] not in registered_storage_classes:
         logger.info(f"Volume creation for storage class that we are not monitoring. ({spec['storageClassName']})")
         return
+    
+    sc_params = validate_pvc_spec(spec, meta)
 
     volume_node = meta.annotations[Constants.NODE_SELECTOR_ANNOTATION_KEY]
     current_node_name = config.get().current_node_name
@@ -163,9 +170,10 @@ def create_volume(meta, spec, **kwargs):
 
     desired_volume_size = spec["resources"].get("limits", {}).get("storage", spec["resources"].get("requests", {}).get("storage"))
     access_modes = spec["accessModes"]
-    volume_type = storage_class_params["type"]
+    volume_type = sc_params["type"]
     pv_name = f"pvc-{meta.uid}"
     fs_type = meta.annotations.get(Constants.FILESYSTEM_ANNOTATION_KEY, "xfs")
+    mirror_disk = Constants.MIRROR_ANNOTATION_KEY in meta.annotations
 
     if not desired_volume_size:
         raise kopf.PermanentError("No volume size provided")
@@ -183,8 +191,7 @@ def create_volume(meta, spec, **kwargs):
             raise kopf.PermanentError(f"Cannot use storage path outside of the shared NFS root! (storage path '{storage_path}')")
 
         volume_directory = f"{config.get().shared_nfs_root}/{storage_path}"
-        with suppress(FileExistsError):
-            os.mkdir(volume_directory)
+        os.makedirs(volume_directory, exist_ok=True)
 
         # create the pv object using the main nfs share and the subpath for this volume
         nfs.create_persistent_volume(pv_name, current_node_name, access_modes, desired_volume_size, config.get().current_node_ip, volume_directory, spec["storageClassName"], spec["volumeMode"])
@@ -196,7 +203,7 @@ def create_volume(meta, spec, **kwargs):
 
         if volume_type == Constants.VOLUME_TYPE_ISCSI:
             # provision lvm volume
-            lvm.create_volume(lvm_group, pv_name, fs_type, desired_volume_size)
+            lvm.create_volume(lvm_group, pv_name, fs_type, mirror_disk, desired_volume_size)
 
             # setup iscsi exports using rtstlib-fb
             iscsi_lun = iscsi.export_disk(lvm_group, pv_name)
@@ -210,7 +217,7 @@ def create_volume(meta, spec, **kwargs):
             mount_point = f"/srv/nfs/{pv_name}"
 
             # provision lvm volume then locally mount it where NFS can access it and the set up a NFS share
-            lvm.create_volume(lvm_group, pv_name, fs_type, desired_volume_size, mount_point)
+            lvm.create_volume(lvm_group, pv_name, fs_type, mirror_disk, desired_volume_size, mount_point)
 
             # export the share
             nfs.export_share(mount_point, config.get().nfs_access_cidr)
@@ -221,41 +228,55 @@ def create_volume(meta, spec, **kwargs):
     logger.info(f"Successfully provisioned volume for claim {meta.name}")
 
 @kopf.on.update("persistentvolumeclaim")
-def update_volume_claim(spec, meta, old, new, diff, **kwargs):
-    validate_pvc_spec(spec, meta, update=True)
+def update_volume_claim(spec: Spec, meta: Meta, old: BodyEssence, new: BodyEssence, **kwargs):
+    if spec["storageClassName"] not in registered_storage_classes:
+        logger.info(f"Volume update for storage class that we are not monitoring. ({spec['storageClassName']})")
+        return
+    
+    sc_params = validate_pvc_spec(spec, meta, update=True)
 
-    old_volume_size = old["resources"].get("limits", {}).get("storage", spec["resources"].get("requests", {}).get("storage"))
-    new_volume_size = new["resources"].get("limits", {}).get("storage", spec["resources"].get("requests", {}).get("storage"))
+    old_volume_size = old.get("spec").get("resources", {}).get("limits", {}).get("storage", spec["resources"].get("requests", {}).get("storage"))
+    new_volume_size = new.get("spec").get("resources", {}).get("limits", {}).get("storage", spec["resources"].get("requests", {}).get("storage"))
 
     lvm_group = config.get().lvm_group
     pv_name = f"pvc-{meta.uid}"
 
     if old_volume_size != new_volume_size:
+        if not sc_params["allow_volume_expansion"]:
+            raise kopf.PermanentError(f"Cannot resize Volume. The storageclass {spec['storageClassName']} does not allow it.")
+        
         lvm.resize_volume(lvm_group, pv_name, old_volume_size, new_volume_size)
 
 
 @kopf.on.delete("persistentvolumeclaim")
-def delete_volume_claim(spec, meta, **kwargs):
-    storage_class = spec["storageClassName"]
-    sc_params = get_storage_class_params(storage_class)
+def delete_volume_claim(spec: Spec, meta: Meta, **kwargs):
+    if spec["storageClassName"] not in registered_storage_classes:
+        logger.info(f"Volume deletion for storage class that we are not monitoring. ({spec['storageClassName']})")
+        return
+    
+    sc_params = get_storage_class_params(spec["storageClassName"])
+    if "volumeName" not in spec:
+        logger.info(f"Deleting a PVC that never provisioned '{meta.name}")
+        return
+    
     volume_name = spec["volumeName"]
     retain_volume = sc_params["reclaim_policy"] == "Retain"
 
     if retain_volume:
-        logger.info(f"Retaining PV after deleting PVC '{meta['name']}")
+        logger.info(f"Retaining PV after deleting PVC '{meta.name}")
     else:
-        logger.info(f"Deleting PV after deleting PVC '{meta['name']}")
+        logger.info(f"Deleting PV after deleting PVC '{meta.name}")
 
         core_api = kubernetes.client.CoreV1Api()
         core_api.delete_persistent_volume(volume_name)
 
 
 @kopf.on.delete("persistentvolume")
-def delete_volume(spec, meta, **kwargs):
+def delete_volume(spec: Spec, meta: Meta, **kwargs):
     storage_class = spec["storageClassName"]
     sc_params = get_storage_class_params(storage_class)
     volume_type = sc_params["type"]
-    pv_name = meta["name"]
+    pv_name = meta.name
     lvm_group = config.get().lvm_group
 
     if volume_type == Constants.VOLUME_TYPE_SHARED:
